# SWWAE example: changed conv ksize and switched to ELU activations