@fchollet 2017-03-11
We won't merge new PRs in master. Consider sending the PR to `keras-2` instead.