@antonmbk 2017-03-12
Closing as these changes will be proposed instead for Keras 2, not master.